back-end/requirements.txt
asgiref==3.6.0
Django==3.2.18
djangorestframework==3.14.0
mariadb==1.1.6
mysqlclient==1.4.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.5.0
